Refactor Events and include loading streams from wen-evens
I'm still not happy with how this projects looks like but its getting better.

Split Events modules to Events, EventsStatic and Model.Event. Events holds the runtime parts whereas Model.Events the code that is required to work with instances of 'Event'. The application loads streams.json on startup and merges it with the static streams that have been compiled in.